requirements.txt
# TJRJ Precatórios Scraper Dependencies

# Core dependencies
playwright>=1.40.0
pandas>=2.0.0
pydantic>=2.8.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Development dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.1.0
mypy>=1.8.0
